changed to asFloat to make pizza revenue work

+++ src/metrics.js
@@ -81,7 +81,7 @@ class MetricTracker {
                                     sum: {
                                         dataPoints: [
                                             {
-                                                asInt: metricValue,
+                                                asFloat: metricValue,
                                                 timeUnixNano: Date.now() * 1000000,
                                                 attributes: [],
                                             },
